requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio
argon2-cffi
argon2-cffi-bindings
async-timeout==4.0.2
asynctest==0.13.0
attrs
Babel
backcall
beautifulsoup4==4.8.0
bleach
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.0
certifi
cffi
charset-normalizer
click==7.1.2
colorama
cryptography
debugpy
decorator
defusedxml
entrypoints
eppy==0.5.63
fastjsonschema
# Flask==2.0.1
flit_core
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
httpx
idna
importlib-metadata
importlib-resources
ipywidgets
itsdangerous==2.1.2
jedi
Jinja2
json5
jsonschema
kaleido==0.0.1rc9
ladybug-charts==1.18.8
ladybug-comfort==0.16.29
ladybug-core==0.41.19
ladybug-display==0.9.0
ladybug-geometry==1.26.5
ladybug-pandas==0.4.1
ladybug-radiance==0.2.0
lbt-ladybug==0.27.17
lxml
markdown-it-py==2.2.0
MarkupSafe
matplotlib-inline
mdurl==0.1.2
# mkl-fft==1.3.1
# mkl-random
# mkl-service==2.4.0
multidict==6.0.4
munch==2.5.0
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
notebook
notebook_shim
numpy
packaging
pandas==1.3.5
pandocfilters
parso
pickleshare
Pillow==9.2.0
pkgutil_resolve_name
plotly==5.14.1
ply==3.11
prometheus-client
prompt-toolkit
protobuf==3.20.3
psutil
pyarrow==11.0.0
pycparser
pydeck==0.8.1b0
pydot3k==1.0.17
Pygments==2.15.0
Pympler==1.0.1
# pyOpenSSL
# pyparsing
# pyrsistent
# PySocks
python-dateutil
pytz
pytz-deprecation-shim==0.1.0.post0
pyzmq
requests
ratelimit
rich==13.3.4
scipy==1.7.3
sip
six
smmap==5.0.0
sniffio
soupsieve
streamlit==1.21.0
tenacity==8.2.2
terminado
tinycss2
tinynumpy==1.2.1
toml
tomli
toolz==0.12.0
tornado
tqdm==4.65.0
traitlets
typing_extensions
tzdata==2023.3
tzlocal==4.3
urllib3
validators==0.20.0
watchdog==3.0.0
wcwidth
webencodings==0.5.1
websocket-client
Werkzeug==2.2.3
widgetsnbextension
win-inet-pton
wincertstore==0.2
yarl==1.8.2
zipp
